requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
beautifulsoup4==4.12.2
black==23.9.1
bootstrap-admin==0.4.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.1.1
crispy-bootstrap5==0.7
cryptography==41.0.4
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap-icons==0.8.5
django-bootstrap-v5==1.0.11
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
django-filter==23.3
django-import-export==3.3.1
django-tables2==2.6.0
django-widget-tweaks==1.5.0
django-widgets-improved==1.5.0
django-wkhtmltopdf==3.4.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
flake8==6.1.0
fonttools==4.42.1
greenlet==2.0.2
idna==3.4
japanize-matplotlib==1.1.3
joblib==1.3.2
kiwisolver==1.4.5
line-profiler==4.1.1
Markdown==3.4.4
MarkupPy==1.14
matplotlib==3.8.0
mccabe==0.7.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.26.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pandas==2.1.1
pathspec==0.11.2
Pillow==9.4.0
platformdirs==3.10.0
plotly==5.17.0
psycopg2==2.9.7
pycodestyle==2.11.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyflakes==3.1.0
pyparsing==3.1.1
pyperclip==1.8.2
pyspnego==0.10.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.0.5
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
smbprotocol==1.11.0
soupsieve==2.5
SQLAlchemy==2.0.21
sqlparse==0.4.4
tablib==3.5.0
tenacity==8.2.3
threadpoolctl==3.2.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
xerox==0.4.1
xgboost==2.0.0
xlrd==2.0.1
XlsxWriter==3.1.4
xlwt==1.3.0
